% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{decode_url}
\alias{decode_url}
\title{decodes a vector of URLs}
\usage{
decode_url(urls)
}
\arguments{
\item{urls}{a vector of URLs to decode.}
}
\value{
a character vector containing the decoded versions of urls.
}
\description{
Takes a vector of URLs and consistently decodes them, in
a vectorised way and handling out-of-range characters (unlike \code{\link{URLdecode}})).
}
\examples{
decode_url("https://en.wikipedia.org/wiki/File:Vice_City_Public_Radio_\%28logo\%29.jpg")

\dontrun{
#A demonstrator of the contrasting behaviours around out-of-range characters
URLdecode("\%gIL")
decode_url("\%gIL")
}
}

