% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ip.R
\name{ip_to_numeric}
\alias{ip_to_numeric}
\title{convert IP addresses to their numeric representation}
\usage{
ip_to_numeric(ip_addresses)
}
\arguments{
\item{ip_addresses}{a vector of human-readable IP addresses}
}
\value{
the numeric representation of \code{ip_addresses}, with
invalid IPv4 IPs represented with a 0.
}
\description{
\code{ip_to_numeric} takes IP addresses stored
in their human-readable representation ("192.168.0.1")
and converts it to a numeric representation (3232235521).

It's fully vectorised (yay!) and C++ based (woo!) but
doesn't yet support IPv6 addresses for a variety of
reasons (awww). Invalid IP addresses, including
IPv6 addresses, will have 0 returned.
}
\examples{
#Convert your local, internal IP to its numeric
#representation.
ip_to_numeric("192.168.0.1")
}

