% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{normalise_ips}
\alias{normalise_ips}
\title{Take vectors of IPs and X-Forwarded-For headers and produce single, normalised
IP addresses.}
\usage{
normalise_ips(ip_addresses, x_forwarded_fors)
}
\arguments{
\item{ip_addresses}{a vector of IP addresses}

\item{x_forwarded_fors}{an equally-sized vector of X-Forwarded-For header
contents.}
}
\value{
a vector of IP addresses, incorporating the XFF header value
where appropriate
}
\description{
\code{normalise_ips} takes IP addresses and x_forwarded_for
values and, in the event that x_forwarded_for is non-null, attempts to
extract the "real" IP closest to the client.
}

