% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/agents.R
\name{parse_r_agent}
\alias{parse_r_agent}
\title{parse an R user agent}
\usage{
parse_r_agent(agents)
}
\arguments{
\item{agents}{a vector of user agents}
}
\value{
a data.frame containing "r_version" (the version of R the user agent refers
to), "architecture" (the processor architecture used) and "platform" (windows,
linux or apple). If any of these cannot be identified, they will be represented
by an NA value instead.
}
\description{
takes a vector of R user agents and parses them, extracting R version,
architecture and platform.
}

