% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bro.R
\name{read_bro}
\alias{read_bro}
\title{Read bro logfiles}
\usage{
read_bro(file)
}
\arguments{
\item{file}{the path to the logfile you want to read in}
}
\description{
This function reads logfiles exported from 'bro', automatically
detecting and handling different types of log and returning them as data.frames.
The full range of bro files are not yet supported, but more will be added over time.
}
\examples{
# Read an FTP log
data <- read_bro(system.file("extdata/ftp.log", package = "webreadr"))

}
\seealso{
\code{\link{read_clf}}, \code{\link{read_squid}} and other readers for
different log formats.
}
