% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_elb}
\alias{read_elb}
\title{Read AWS ELB Access Logs}
\usage{
read_elb(file)
}
\arguments{
\item{file}{the full path to the ELB log file you want to read.}
}
\description{
\code{read_elb} provides a reader for AWS ELB access logs.
}
\examples{
# Using the inbuilt testing dataset
data <- read_elb(system.file("extdata/log.elb", package = "webreadr"))

}
