% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_iis}
\alias{read_iis}
\title{Read Microsoft IIS Logs}
\usage{
read_iis(file)
}
\arguments{
\item{file}{the full path to the IIS log file you want to read.}
}
\description{
\code{read_iis} provides a reader for Microsoft IIS access logs.
}
\examples{
# Using the inbuilt testing dataset
iis_data <- read_iis(system.file(file.path("extdata", "iis.log"), package = "webreadr"))

}
