% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readers.R
\name{read_s3}
\alias{read_s3}
\title{Read Amazon S3 Access Logs}
\usage{
read_s3(file)
}
\arguments{
\item{file}{the full path to the S3 file you want to read.}
}
\description{
\code{read_s3} provides a reader for Amazon's S3 service's access logs, described
\href{http://docs.aws.amazon.com/AmazonS3/latest/dev/LogFormat.html}{here}.
}
\examples{
# Using the inbuilt testing dataset
s3_data <- read_s3(system.file("extdata/s3.log", package = "webreadr"))

}
\seealso{
\code{\link{read_aws}} for reading Amazon Web Services (AWS) access log files.
}

