% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/splitters.R
\name{split_clf_requests}
\alias{split_clf_requests}
\title{split requests from a CLF-formatted file}
\usage{
split_clf_requests(requests)
}
\arguments{
\item{requests}{the "request" field from a CLF-formatted file, read in with
\code{\link{read_clf}} or \code{\link{read_combined}}.}
}
\value{
a data.frame of three columns - "method", "asset" and "protocol" -
representing, respectively, the HTTP method used ("GET"), the asset requested
("/favicon.ico") and the protocol used ("HTTP/1.0").
}
\description{
CLF (Combined/Common Log Format) files store the HTTP method, protocol
and asset requested in the same field. \code{split_clf} takes this field as a vector
and returns a data.frame containing these elements in distinct columns.
}
\examples{
#Grab CLF data and split out the request.
data <- read_combined(system.file("extdata/combined_log.clf", package = "webtools"))
requests <- split_clf_requests(data$request)
}
\seealso{
\code{\link{read_clf}} and \code{\link{read_combined}} for reading
in these files.
}

