% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webreadr.R
\name{webreadr}
\alias{webreadr}
\title{A package for reading various common forms of request log}
\description{
see the \href{https://github.com/Ironholds/webreadr/blob/master/vignettes/Introduction.Rmd}{introductory vignette}
for more details!
}

