% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/webtools.R
\name{webtools}
\alias{webtools}
\title{A package for reading, formatting and cleaning various common forms of web request log.}
\description{
see the \href{https://github.com/Ironholds/webtools/blob/master/vignettes/Introduction.Rmd}{introductory vignette}
for more details!
}

